Add project array files
Project Description

You are given an interface for a type of list. The list works like this:

entries begin at index 0
the size of the list is limited ("capped"); when the size is reached, no more entries can be added
entries can only be added to and removed from the front or back of the list
entries can be accessed in any position
Write a class that implements this interface. The class uses arrays to implement the list.